services/langgraph_pipeline/requirements.txt
langgraph==0.0.19
neo4j==5.20.0
kafka-python==2.0.2
psycopg[binary]==3.1.19
pgvector==0.2.4
networkx==3.2.1
